data_process/requirements.txt
pandas
rich
pyyaml
matplotlib
seaborn